Relocate and rename `sbi_posterior.py`.
